Fix ident for advisory databases
